Move str tests to tests/types/test_str.py
Also split the str cases out of the `test_default_validators`
parameterization and replace `constr()` usages with the explicit
annotated forms using `StringConstraints`.
